\name{ppc.plot}
\alias{ppc.plot}
\encoding{UTF-8}
\title{
Plot results from posterior predictive check
}
\description{
Function plots results from posterior predictive check in AHM1 section 6.8 for a fitted model object with JAGS.
}
\usage{
ppc.plot(fm)
}
\arguments{
  \item{fm}{
The fitted model object
}
}
\value{
None, used for its plotting side effect.
}
\references{
Kéry & Royle (2016) \emph{Applied Hierarchical Modeling in Ecology} Section 6.8

}
\author{
 Marc Kéry,
}
\examples{
## We would need a fitted model object to demonstrate this, but that takes a long time to run
}
