\name{predict-methods}
\docType{methods}
\alias{predict-methods}
\alias{predict,unmarkedFitPCount-method}
\encoding{UTF-8}
\title{ Methods for Function predict, signature unmarkedFit in Package `unmarked' }
\description{
This return predicted values from fitted model objects, including SE and confidence intervals for ZIP models.
}
\section{Methods}{
\describe{
\item{\code{signature(object = "unmarkedFit")}}{
"type" must be either `state' or `det'.
}

}}
\keyword{methods}