\name{AHMbook-defunct}
\alias{AHMbook-defunct}
\alias{plot.Nmix.resi}
\alias{image.scale}
\encoding{UTF-8}
\title{
Defunct Functions in Package AHMbook
}
\description{
Functions in the AHMbook package with have been replaced.
}
\usage{
plot.Nmix.resi(...)
image.scale(...)
}
\arguments{
  \item{...}{
arguments passed to the old function.
}
}

\details{
 
\code{plot.Nmix.resi} has been renamed \code{plot_Nmix_resi}, as it was being treated by R as a method for \code{plot} generic.

\code{image.scale} has been renamed \code{image_scale}, as it was being treated by R as a method for \code{image} generic.
}

