\name{AHMbook-package}
\alias{AHMbook-package}
\alias{AHMbook}
\docType{package}
\title{
Functions and data for the Book 'Applied Hierarchical Modeling in Ecology'
}
\description{
Provides functions and data sets needed to run the code provided in the two-volume publication, \emph{Applied Hierarchical Modeling in Ecology: analysis of distribution, abundance and species rishness in R and BUGS} by Marc Kéry and Andy Royle, Academic Press (Vol 1, 2016; Vol 2, in press).
}
\details{


The functions are listed by chapter below.
}

\section{Chapter 1}{
 
  \tabular{ll}{
  \code{\link{sim.fn}} \tab simulation of a Poisson process} }
  
\section{Chapter 4}{
 
  \tabular{ll}{
  \code{\link{data.fn}} \tab ...} }
  
\section{Chapter 6}{
 
  \tabular{ll}{
  \code{\link{data.fn}} \tab ... \cr
  \code{\link{ppc.plot}} \tab ... \cr
  \code{\link{plot.Nmix.resi}} \tab ... \cr
  \code{\link{map.Nmix.resi}} \tab ... \cr
  \code{\link{simpleNmix}} \tab ... \cr
  \code{\link{playRN}} \tab ...   } }
  
\section{Chapter 7}{
 
  \tabular{ll}{
  \code{\link{fitstats}}, \code{\link{fitstats2}} \tab ...\cr
  \code{\link{piFun}} \tab ...} }
  
\section{Chapter 8}{
 
  \tabular{ll}{
  \code{\link{sim.ldata}} \tab ...\cr
  \code{\link{sim.pdata}} \tab ...\cr
  \code{\link{simHDS}} \tab ...} }
  
\section{Chapter 9}{
 
  \tabular{ll}{
  \code{\link{simHDSg}} \tab ...\cr
  \code{\link{simHDStr}} \tab ...\cr
  \code{\link{simHDSopen}} \tab Simulate open hierarchical distance sampling data \cr
  \code{\link{issj.sim}} \tab ...\cr
  \code{\link{sim.spatialDS}} \tab ...\cr
  \code{\link{simHDStr}} \tab ...\cr
  \code{\link{pcount.spHDS}} \tab ...} }

\section{Chapter 10}{
 
  \tabular{ll}{
  \code{\link{simOcc}} \tab ...\cr
  \code{\link{sim3Occ}} \tab ...\cr
  \code{\link{simOccttd}} \tab ...\cr
  \code{\link{wigglyOcc}} \tab ...\cr
  \code{\link{spline.prep}} \tab ...} }

\section{Chapter 11}{
 
  \tabular{ll}{
  \code{\link{simComm}} \tab Simulate community occupancy or abundance data } }
  
\section{Chapter 15}{
 
  \tabular{ll}{
  \code{\link{simDynocc.fn}} \tab ...} }
  
\section{Chapter 20}{
 
  \tabular{ll}{
  \code{\link{spatial.exp}} \tab ...} }
  
%\section{Chapter 21}{
% 
%  \tabular{ll}{
%  \code{\link{data.fn}} \tab ...} } %% need to rename this
  
  
\section{DATA SETS}{

  \tabular{ll}{
  \code{\link{jay}} \tab ... }
}

\section{UTILITY FUNCTIONS}{

  \tabular{ll}{
  \code{\link{e2dist}} \tab ... }
}

\author{
Marc Kéry and Andy Royle

}

\keyword{ package }
