\name{ppc.plot}
\alias{ppc.plot}
\encoding{UTF-8}
\title{
Plot results from posterior predictive check
}
\description{
Function plots results from posterior predictive check in AHM1 section 6.8 for a fitted model object with JAGS.
}
\usage{
ppc.plot(fm)
}
\arguments{
  \item{fm}{
The fitted model object
}
}
\value{
None, used for its plotting side effect.
}
\references{
Kéry, M. & Royle, J.A. (2016) \emph{Applied Hierarchical Modeling in Ecology} AHM1 - 6.8.

}
\author{
Marc Kéry & Andy Royle
}
